it("'instance'.update()",function(done){ //stopping point
  Update DB Records Test Suite
{ AssertionError [ERR_ASSERTION]: The expression evaluated to a falsy value:

  assert(data.userState === "active")

    at C:\Users\Dell\Documents\GitHub\WebNinja_MongoDB-Tutorial_Jan2017\test\updateDB_test.js:32:17
    at process._tickCallback (internal/process/next_tick.js:68:7)
  generatedMessage: true,
  name: 'AssertionError [ERR_ASSERTION]',
  code: 'ERR_ASSERTION',
  actual: false,
  expected: true,
  operator: '==' }
(node:14272) UnhandledPromiseRejectionWarning: ReferenceError: data is not defined
    at C:\Users\Dell\Documents\GitHub\WebNinja_MongoDB-Tutorial_Jan2017\test\updateDB_test.js:37:29
    at process._tickCallback (internal/process/next_tick.js:68:7)
(node:14272) UnhandledPromiseRejectionWarning: Unhandled promise rejection. This error originated either by throwing inside of an async function without a catch block, or by rejecting a promise which was not handled with .catch(). (rejection id: 1)
(node:14272) [DEP0018] DeprecationWarning: Unhandled promise rejections are deprecated. In the future, promise rejections that are not handled will terminate the Node.js process with a non-zero exit code.
    1) 'instance'.update()
    √ 'Model'.update()
    √ 'Model'.findOneAndUpdate